Consolidate the display-purity tests on revert-sweep evidence
Trimmed by measurement, not by feel: each of the 18 production hunks on this
branch was reverted ON ITS OWN and the whole javatools suite run, recording which
tests went red. 8 files / 1491 lines -> 5 files / 1394 lines, with no assertion
weakened and nothing deleted that uniquely catches anything.

WHAT THE SWEEP FOUND

Nine reverts turn something red, and in every case the catcher is the ONLY test
that catches it - so no surviving test is redundant:

  ParameterizedTypeConstant.getValueString -> DisplayPurityTest x2
  TerminalTypeConstant.getValueString      -> DisplayPurityTest
  ParamInfo.toString                       -> DisplayPurityTest
  Annotation display                       -> DisplayPurityTest
  BinaryAST.toString                       -> DisplayPurityTest
  MethodStructure.getDescription           -> DisplayPurityTest (merged in here)
  MethodInfo.toString                      -> DisplayPurityRuntimeTest
  ObjectHandle.toString                    -> ObjectHandleDisplayPurityTest
  DeferredArrayHandle.toString             -> DisplayPurityCensusTest

Nine reverts turn NOTHING red: PropertyBody.toString,
Component.Contribution.toString, the ensureCaches() avoidance in
TypeInfoReal.toString, ExceptionHandle.toString, Proxy.ProxyHandle.toString,
xRTType.TypeHandle.toString, xClass.ClassHandle.toString,
xFuture.FutureHandle.toString and xString.StringHandle.toString. Each mutates by
a mechanism a warmed pool cannot show - an already-interned canonical constant, a
private lazy cache, an allocation, a memoized field - so they are
inspection-verified only. That list is now recorded in the census's own javadoc,
where a reader will find it; closing any of them needs an observable, not another
assertion over the same population.

A BUG THE SWEEP FOUND IN MY OWN TEST

ObjectHandleDisplayPurityTest.renderingAHandleNeverAsksItsCompositionForATypeConstant
did NOT catch its revert. The pre-fix body was

    m_fMutable || clz.getType().isImmutable()

so on a mutable handle the || short-circuits and getType() is never reached: the
assertion held against the unfixed code and proved nothing. Only its sibling
caught the revert, by accident, because that one sets m_fMutable = false. The
test now freezes the handle first, and both halves catch. This is the second time
a purity assertion here has been green for the wrong reason; the comment says so.

WHAT MOVED

  MethodDisplayPurityTest (55)   -> merged into DisplayPurityTest; same fixture
                                    (cold FileStructure pool), same kind of
                                    assertion, still the unique catcher.
  TypeInfoDisplayPurityTest (84) -> merged into DisplayPurityRuntimeTest; both
                                    need the XDK-backed container.
  HandlePopulation (130)         -> folded into DisplayPurityCensusTest, its only
                                    consumer. That also settles the two-fixtures
                                    smell: DisplayPurityFixture is now the single
                                    fixture, and it does one thing - locate the
                                    built XDK and start a runtime.

DELETED OUTRIGHT: DisplayPurityRuntimeTest.renderingTheTypeSystemDoesNotGrowThe-
ConstantPool. It was the unique catcher for nothing, and it is the specific test
whose warm-pool design once passed against impure code. Its job is done properly
elsewhere - cold-pool assertions in DisplayPurityTest, and the per-object,
first-render version in the census. Its negative control is not lost; the
surviving cold-pool tests each carry one.

KEPT DELIBERATELY: every negative control, the no-warm-up comment, and the "what
this proves and does not prove" javadoc.
